requirements.txt
torch
transformers
peft
flask
Pillow
